+++ R/transform_runner.R
@@ -20,19 +20,20 @@ run_transform <- function(con, transform, path, extracted_data,
     if (is.null(extracted_data)) {
       stop("Cannot run transform as no data has been extracted.")
     }
-    transformed_data <- transform(extracted_data)
-    verify_data(con, transformed_data, mode)
-    if (!is.null(transform_test)) {
-      message(sprintf("Running transform tests %s", transform_test))
-      test_path <- file.path(path, transform_test)
-      test_results <-
-        run_transform_tests(test_path, transformed_data, extracted_data, con)
-      if (!all_passed(test_results)) {
-        stop("Not all transform tests passed. Fix tests before proceeding.")
-      } else {
-        message("All transform tests passed.")
+    withr::with_dir(path, {
+      transformed_data <- transform(extracted_data)
+      verify_data(con, transformed_data, mode)
+      if (!is.null(transform_test)) {
+        message(sprintf("Running transform tests %s", transform_test))
+        test_results <-
+          run_transform_tests(transform_test, transformed_data, extracted_data, con)
+        if (!all_passed(test_results)) {
+          stop("Not all transform tests passed. Fix tests before proceeding.")
+        } else {
+          message("All transform tests passed.")
+        }
       }
-    }
+    })
     transformed_data
   }