+++ tests/testthat/example_relative_path/R/extract.R
@@ -1,0 +1,5 @@
+extract <- function(con) {
+  raw_data <- list()
+  raw_data$people <- read.csv("data/people.csv", stringsAsFactors = FALSE)
+  raw_data
+}