+++ tests/testthat/helper-dettl-config.R
@@ -16,11 +16,12 @@ setup_config <- function(db_driver = "RSQLite::SQLite",
   path
 }
 
-setup_dettl_config <- function(load = "func: load") {
+setup_dettl_config <- function(load = "func: load", dettl = "") {
   path <- setup_dettl("example_load_template", "dettl_config.yml")
   filename <- file.path(path, "example_load_template", "dettl.yml")
   cfg <- readLines(filename)
   cfg <- gsub("<load>", load, cfg, fixed = TRUE)
+  cfg <- gsub("<dettl>", dettl, cfg, fixed = TRUE)
   cfg <- unlist(strsplit(cfg, "\n"))
   writeLines(cfg, filename)
   file.path(path, "example_load_template")