+++ NEWS.md
@@ -1,3 +1,8 @@
+# dettl 0.0.13
+
+* Set working directory to transform & load stages to be the import directory (working dir now consistent over extract, transform and load)
+* Allow running the load stage outside of a transaction (if dry_run = TRUE then it will still be run inside a transaction)
+
 # dettl 0.0.12
 
 * Remove path arg from extract stage - this will break previous imports